move persistent-statetul-service to services/core
