+++ app/services/core/persistent-stateful-service.ts
@@ -1,6 +1,6 @@
 import { merge } from 'lodash';
-import { StatefulService } from './core/stateful-service';
-import Utils from './utils';
+import { StatefulService } from './stateful-service';
+import Utils from '../utils';
 
 // Extends StatefulService with code that will persist the
 // state across executions of the application.
