+++ app/services/customization/customization.ts
@@ -1,5 +1,5 @@
 import { Subject } from 'rxjs';
-import { PersistentStatefulService } from '../persistent-stateful-service';
+import { PersistentStatefulService } from '../core/persistent-stateful-service';
 import { mutation } from '../core/stateful-service';
 import {
   ICustomizationServiceApi,
