+++ app/services/dismissables.ts
@@ -1,4 +1,4 @@
-import { PersistentStatefulService } from 'services/persistent-stateful-service';
+import { PersistentStatefulService } from 'services/core/persistent-stateful-service';
 import { mutation } from './core/stateful-service';
 import Vue from 'vue';
