+++ app/services/nicolive-program/nicolive-comment-local-filter.ts
@@ -1,4 +1,4 @@
-import { PersistentStatefulService } from 'services/persistent-stateful-service';
+import { PersistentStatefulService } from 'services/core/persistent-stateful-service';
 import { mutation } from 'services/core/stateful-service';
 import { isAnonymous, getScore } from './ChatMessage/util';
 import { WrappedChat } from './WrappedChat';
