+++ app/services/patch-notes/index.ts
@@ -1,4 +1,4 @@
-import { PersistentStatefulService } from 'services/persistent-stateful-service';
+import { PersistentStatefulService } from 'services/core/persistent-stateful-service';
 import { mutation } from 'services/core/stateful-service';
 import electron from 'electron';
 import Util from 'services/utils';
