+++ app/services/user.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import URI from 'urijs';
-import { PersistentStatefulService } from 'services/persistent-stateful-service';
+import { PersistentStatefulService } from 'services/core/persistent-stateful-service';
 import { Inject } from 'services/core/injector';
 import { mutation } from 'services/core/stateful-service';
 import electron from 'electron';
